Removing XUnit tests; renaming MsTest tests
