--- microsoft-azure-servicelayer/Microsoft.WindowsAzure.ServiceLayer.UnitTests.MsTest/.gitignore
@@ -1,3 +1,0 @@
-*.suo
-bin/
-obj/
