--- microsoft-azure-servicelayer/Microsoft.WindowsAzure.ServiceLayer.UnitTests.MsTest/HttpTests/TestHttpHandler.cs
@@ -1,57 +1,0 @@
-﻿//
-// Copyright 2012 Microsoft Corporation
-// 
-// Licensed under the Apache License, Version 2.0 (the "License");
-//  you may not use this file except in compliance with the License.
-//  You may obtain a copy of the License at
-//    http://www.apache.org/licenses/LICENSE-2.0
-// 
-//  Unless required by applicable law or agreed to in writing, software
-//  distributed under the License is distributed on an "AS IS" BASIS,
-//  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-//  See the License for the specific language governing permissions and
-//  limitations under the License.
-//
-
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using Microsoft.WindowsAzure.ServiceLayer.Http;
-
-namespace Microsoft.WindowsAzure.ServiceLayer.UnitTests.MsTest.HttpTests
-{
-    /// <summary>
-    /// HTTP handler for testing pipelining.
-    /// </summary>
-    internal class TestHttpHandler: IHttpHandler
-    {
-        internal int BeforeCount { get; set; }
-        internal int AfterCount { get; set; }
-        internal IHttpHandler NextHandler { get; private set; }
-
-        /// <summary>
-        /// Initializes the object.
-        /// </summary>
-        internal TestHttpHandler(IHttpHandler nextHandler)
-        {
-            BeforeCount = 0;
-            AfterCount = 0;
-            NextHandler = nextHandler;
-        }
-
-        /// <summary>
-        /// Processes the request.
-        /// </summary>
-        /// <param name="request">Original HTTP request.</param>
-        /// <returns>HTTP response.</returns>
-        HttpResponse IHttpHandler.ProcessRequest(HttpRequest request)
-        {
-            BeforeCount++;
-            HttpResponse response = NextHandler.ProcessRequest(request);
-            AfterCount++;
-            return response;
-        }
-    }
-}
