+++ microsoft-azure-servicelayer/Microsoft.WindowsAzure.ServiceLayer.UnitTests/.gitignore
@@ -1,4 +1,3 @@
 *.suo
 bin/
 obj/
-
