+++ microsoft-azure-servicelayer/Microsoft.WindowsAzure.ServiceLayer.UnitTests/Images/UnitTestStoreLogo.png
